chore: Add production build on server
Add script that creates a production build which we can
deploy on the server since babel-node is not meant to be
used in production.

Also, moved all testing dependencies and utils to
dev-dependencies so they don't come up on the production
server.

Also, moved testUtils.js in __tests__ folder so it
doesn't come up in production build.

resolves #77